Fixed warning about comparison of integer expressions of different signedness
size_ is a stream_offset which is a boost::intmax_t which is signed
max_length is a mapped_file_source::size_type which is a std::size_t
which is unsigned

It is not expected that size_ be negative here so we cast it to the
unsigned integer. However we prefer to be safe by using
boost::numeric_cast so that in the event there is actually an overflow
it will be handled correctly with an exception.

Fixes #63