Add multithreaded_test to lzma_test
Verifies that various number of threads arrive at the
same compressed data. While it's doing that, it verifies
that various constructors compile and function properly.

It does not test that threading is actually occurring. That
would be challenging. And the correct behavior is not to
use multiple threads, if liblzma doesn't have that
capability.