Use multithreaded versions of liblzma functions
The previous constructor for the filter continues to work as
before. But you can now specify a threads value in the
lzma_mt parameter struct. Default value is 1, meaning
single-threaded.

If you ask for 0 threads, it will automatically detect the number
of cores and use that. To accomplish this, we leverage this capability
in liblzma already.

Liblzma versions before 5.2.0 didn't have multithreading
capabilities and didn't have the necessary functions. So we detect
the liblzma version using the LZMA_VERSION macro, and fallback to
the previous single-threaded implementation when necessary.

If liblzma was built with --disable-threads or --enable-threads=no,
then liblzma will not have the required multithreaded functions,
even if the version is sufficiently recent. We have no way to
detect this case at compile-time (no macro in liblzma's headers)
and it will lead to link-time errors. So we provide the
capability to disable the new multithreaded implementation by
defining BOOST_IOSTREAMS_LZMA_NO_MULTITHREADED.