Disable msvc warning that can occur under some versions+settings (#161)
This warning has been observed under more than one version. One
specific example, as described in issue #138, is on
indirect_streambuf.hpp(440) when Whole Program Optimization is off.
In this case, the line before is seen by the optimizer as always
throwing for particular instantiations of the template,
and therefore the line in question is flagged as not reachable.

MSVC team has stated this is is expected behavior.

This is a simple and hopefully good-enough fix.